requirements.txt
kafka-python
pandas
python-dotenv
confluent-kafka

